♻️ Refactor reader routes and check if user can read
